+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - New function to list users that the user has disabled retweets from (`friends_no_retweets`)
 - New function to look up friendship status for several users (`relation_lookup`)
 - New functions to block/unblock users (`block`, `unblock`, `report_spam`)
+- New functions to mute/unmute users (`mute`, `unmute`)
 
 ### Changed
 - Combined IDLoader and UserLoader into CursorIter (with the same interface)